Conflicts:
	src/mfast/instructions/sequence_instruction.h

+++ src/mfast/instructions/sequence_instruction.h
@@ -95,7 +95,8 @@ namespace mfast
     {
       // This member function is used to query the properties of sequence element.
       // Becuase the sequence element can never be optional, the presence bit should never be encoded.
-      return  /* (optional() ? 1 : 0) | */ segment_pmap_size()  << 1;
+
+      return  /* (optional() ? 1 : 0) | */ static_cast<int>(segment_pmap_size()  << 1);
     }
 
   private:
